+++ frontend/src/assets/styles/font.css
@@ -5,6 +5,9 @@
   font-style: normal;
 }
 
-body {
-  font-family: "MyFont", sans-serif;
+@font-face {
+  font-family: "MyLogoFont";
+  src: url("/assets/font/OneShinhanBold.ttf") format("truetype");
+  font-weight: normal;
+  font-style: normal;
 }
