+++ frontend/src/layouts/ComponentsInMain/PenaltySidebar.tsx
