+++ frontend/src/layouts/MyNavbar/MyNavbar.tsx
@@ -1,30 +1,45 @@
 "use client";
 
-import "bootstrap/dist/css/bootstrap.min.css";
 import Container from "react-bootstrap/Container";
 import Nav from "react-bootstrap/Nav";
 import Navbar from "react-bootstrap/Navbar";
-import Image from "next/image";
 import Link from "next/link";
+import "../../assets/styles/font.css";
 
 export default function AppNavbar() {
   return (
-    <Navbar collapseOnSelect expand="lg" className="bg-body-tertiary">
-      <Container>
-        <Navbar.Brand as={Link} href="/">
-          <Image
-            src="/assets/신한투자증권.jpg"
-            alt="로고"
-            width={120}
-            height={40}
-            style={{ objectFit: "contain" }}
+    <Navbar collapseOnSelect expand="lg" style={{}}>
+      <Container className="d-flex justify-content-between align-items-center">
+        <Navbar.Brand
+          as={Link}
+          href="/"
+          style={{
+            display: "flex",
+            alignItems: "center",
+          }}
+        >
+          <img
+            src="/assets/신한투자증권.png"
+            alt="logo"
+            style={{ width: "32px", height: "32px", marginRight: "10px" }}
           />
+          <div
+            style={{
+              fontFamily: "MyLogoFont",
+              color: "#0046ff",
+              fontSize: "20px",
+            }}
+          >
+            PDA_LOG
+          </div>
         </Navbar.Brand>
 
         <Navbar.Toggle aria-controls="responsive-navbar-nav" />
-
-        <Navbar.Collapse id="responsive-navbar-nav">
-          <Nav className="ms-3 d-flex gap-2">
+        <Navbar.Collapse
+          id="responsive-navbar-nav"
+          className="justify-content-end"
+        >
+          <Nav className="d-flex gap-2">
             <Nav.Link
               as={Link}
               href="/login"
